Fix the issue with sending logs to grafana
Application logs are not being sent to grafana.
That's because the grafana servers were migrated to docker, but we were
using grafana IP addresses instead of dns endpoints, so the logs no
longer were being delivered.
We have changed the configuration and added the dns hostnames of
grafana.

Fixes issue: SAASOPS-12201
Related issues: SAASOPS-12123, SAASOPS-12149